Works with all testcases so far
[Makefile] Updated the Makfile in root directory.

[max-slicing] RegionManager is working.

[max-slicing] An original instruction may have multiple cloned copies.
Modified CloneInfoManager and related code to support this.

[max-slicing] If a cloned BB doesn't have a terminator due to incomplete
execution, link it to the unreachable BB in its function, so that we can
later recognize it by using MaxSlicing::is_unreachable.

[max-slicing] Fixed a bug: MaxSlicing::is_unreachable shouldn't just
look at the name. The name might be modified by optimization when
combining BBs.

[int] Fixed a bug in contains_only_consts. A constant should be always
considered as a constant, no matter whether it appears in CC.constants.

[simplifier] Should run LoopSimplify before AggressiveLoopUnroll.

[tests] Fixed a bug in int-test: The local thread IDs in FFT may be
highly optimized in the simplified program. Therefore, search for
StoreInst's instead of Add's.